Generalize inference of repository configuration
Previously ActiveRecord was required to automatically infer the database
configuration.  This has been generalized and now also supports
DataMapper, as well as attempting to fall back to the environment
variable DATABASE_URL if available.
